Add basic HEVC / H.265 recording support.
I have followed the T-REC-H.265-201911-I!!PDF-E.pdf documentation to the
best of my understanding. This may need enhanced as new encoder techniques
are encountered.

+++ mythtv/libs/libmythtv/recorders/recorderbase.cpp
@@ -790,7 +790,7 @@ void RecorderBase::ResolutionChange(uint width, uint height, long long frame)
     }
 }
 
-void RecorderBase::FrameRateChange(uint framerate, long long frame)
+void RecorderBase::FrameRateChange(uint framerate, uint64_t frame)
 {
     if (m_curRecording)
     {
